+++ pf_vprintf.c
@@ -49,7 +49,7 @@ int ft_vsnprintf(char *str, size_t size, const char *format, va_list ap)
 			}
 		}
 	}
-	return (p.bytes);
+	return ((int)p.bytes);
 }
 
 //dprintf()'s va_list function.
